Updated the test to control the node heartbeat timer. This should significantly speed up the tests

+++ paxos/acceptor.cc
@@ -112,7 +112,8 @@ void RunServer(const std::string& address, uint8_t node_id,
   while (!stoken.stop_requested()) {
     std::this_thread::sleep_for(300ms);
   }
-  LOG(INFO) << "NODE: [" << node_id << "] Shutting down acceptor service";
+  LOG(INFO) << "NODE: [" << static_cast<uint32_t>(node_id)
+            << "] Shutting down acceptor service";
 }
 
 AcceptorService::AcceptorService(const std::string& address, uint8_t node_id,
